+++ example/lib/main.dart
@@ -35,7 +35,7 @@ class _MyAppState extends State<MyApp> {
     } on PlatformException {
       platformVersion = 'Failed to get platform version.';
     }
-    isTablet = await CoderMatthewsExtensions.isTablet();
+    isTablet = await TabletExtensions.isTablet();
     // If the widget was removed from the tree while the asynchronous platform
     // message was in flight, we want to discard the reply rather than calling
     // setState to update our non-existent appearance.