+++ lib/src/extensions/object_extensions.dart
@@ -1,6 +1,17 @@
 extension ObjectExn on Object? {
   bool get isNull => this == null;
   bool get isNotNull => this != null;
+  bool get isPrimitiveDataType {
+    switch (runtimeType) {
+      case String:
+      case num:
+      case DateTime:
+      case bool:
+        return true;
+      default:
+        return false;
+    }
+  }
 }
 
 extension FutureExn<T> on Future<T?> {