+++ lib/src/helpers/enums.dart
@@ -1,0 +1,1 @@
+enum OrderDirection { desc, asc }