+++ lib/src/helpers/validation_contract.dart
@@ -1,6 +1,6 @@
-import 'object_extensions.dart';
-import 'string_extensions.dart';
-import 'exceptions/app_exception.dart';
+import '../exceptions/exceptions.dart';
+import '../extensions/object_extensions.dart';
+import '../extensions/string_extensions.dart';
 
 class ValidationContract {
   static void requires(bool condition, String title, String message) {
@@ -23,28 +23,23 @@ class ValidationContract {
     if (val.isNotNullOrEmpty) throw AppException(title, message);
   }
 
-  static void requiresWithCallback(
-      bool condition, String title, String message, void Function() callback) {
+  static void requiresWithCallback(bool condition, String title, String message, void Function() callback) {
     if (!condition) throw AppException(title, message, callback);
   }
 
-  static void requireNotNullWithCallback<T>(
-      T? val, String title, String message, void Function() callback) {
+  static void requireNotNullWithCallback<T>(T? val, String title, String message, void Function() callback) {
     if (val.isNull) throw AppException(title, message, callback);
   }
 
-  static void requireNullWithCallback<T>(
-      T? val, String title, String message, void Function() callback) {
+  static void requireNullWithCallback<T>(T? val, String title, String message, void Function() callback) {
     if (val.isNotNull) throw AppException(title, message, callback);
   }
 
-  static void requireNotNullOrEmptyWithCallback(
-      String? val, String title, String message, void Function() callback) {
+  static void requireNotNullOrEmptyWithCallback(String? val, String title, String message, void Function() callback) {
     if (val.isNullOrEmpty) throw AppException(title, message, callback);
   }
 
-  static void requireNullOrEmptyWithCallback(
-      String? val, String title, String message, void Function() callback) {
+  static void requireNullOrEmptyWithCallback(String? val, String title, String message, void Function() callback) {
     if (val.isNotNullOrEmpty) throw AppException(title, message, callback);
   }
 }