+++ test/map_extensions_test.dart
@@ -3,11 +3,7 @@ import 'package:test/test.dart';
 
 void main() {
   test('shouldRemoveNullsFromMap', () {
-    var testMap = <String, dynamic>{
-      'name': 'Tom',
-      'age': null,
-      'location': 'USA'
-    };
+    var testMap = <String, dynamic>{'name': 'Tom', 'age': null, 'location': 'USA'};
     var control = <String, dynamic>{'name': 'Tom', 'location': 'USA'};
     var newMap = testMap.removeNulls;
     expect(control.entries.length, newMap.entries.length);